requirements.txt
llama-index==0.10.30
streamlit==1.33.0
python-dotenv==1.0.1